# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Urban Sanctuary API
  description: |
    Premium Apartment Booking Platform — Douala, Cameroon.

    The Urban Sanctuary API powers a modern apartment booking system with:
    - Multi-method authentication (Email/Password, Phone/OTP, 2FA)
    - Property listing management with image uploads
    - Booking engine with conflict prevention & loyalty points
    - Payment integration (MTN Mobile Money & Orange Money)
    - Role-based access control (Guest, Homeowner, Admin)
    - Review & rating system (verified stays only)
    - Full admin console with audit trail

    ## Authentication
    All authenticated endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens are obtained via `/auth/login` or `/auth/register` and are valid for **15 minutes**.
    Refresh tokens are stored as HttpOnly cookies and are valid for **7 days**.

    ## Response Envelope
    All responses follow a standard JSON envelope:
    ```json
    {
      "success": true,
      "status": 200,
      "message": "...",
      "data": { ... },
      "meta": { "page": 1, "per_page": 20, "total": 100, "total_pages": 5 },
      "request_id": "uuid"
    }
    ```

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Auth endpoints: 5 requests/minute per IP
    - Rate limit headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Rotates the refresh token and issues a new access token.
        The refresh token can be sent via HttpOnly cookie (preferred) or in the request body.
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Only needed if not using HttpOnly cookie
      responses:
        "200":
          description: Token refreshed
          headers:
            Set-Cookie:
              description: New HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/otp/request:
    post:
      tags: [Auth]
      summary: Request OTP
      description: |
        Sends a 6-digit OTP code via SMS or email.
        Codes expire after **5 minutes** and allow a maximum of **3 verification attempts**.
      operationId: requestOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP
      description: Validates a 6-digit OTP code.
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPVerifyRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Enable 2FA
      description: Initiates TOTP 2FA setup. Returns a secret and QR code URL.
      operationId: enable2FA
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 2FA setup initiated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Enable2FAResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/2fa/confirm:
    post:
      tags: [Auth]
      summary: Confirm 2FA
      description: Confirms TOTP 2FA activation with a 6-digit code from the authenticator app.
      operationId: confirm2FA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TOTPCodeRequest"
      responses:
        "200":
          description: 2FA activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Verify 2FA
      description: Validates a TOTP 2FA code during login.
      operationId: verify2FA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TOTPCodeRequest"
      responses:
        "200":
          description: 2FA verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/password/reset-request:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Sends a password reset OTP to the user's email or phone.
      operationId: passwordResetRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Reset instructions sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password
      description: Sets a new password using the reset token.
      operationId: passwordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirm"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ──────────────────────────────────────────────
  # Users
  # ──────────────────────────────────────────────
...
  /properties:
    get:
      tags: [Properties]
      summary: List properties
      description: |
        Returns published properties with support for filtering, searching, sorting, and pagination.
        Only properties with status `published` are returned to unauthenticated users.
      operationId: listProperties
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: neighborhood
          in: query
          schema:
            type: string
            enum: [akwa, bonapriso, bonanjo, kotto, other]
        - name: min_price
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: max_price
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: bedrooms
          in: query
          schema:
            type: integer
            minimum: 1
        - name: max_guests
          in: query
          schema:
            type: integer
            minimum: 1
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, price_per_night]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Full-text search on title and description
          schema:
            type: string
      responses:
        "200":
          description: Properties retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PropertyResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Properties]
      summary: Create a property
      description: Creates a new property listing in `draft` status. Requires `homeowner` or `admin` role.
      operationId: createProperty
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PropertyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
        request_id:
          type: string
          format: uuid
      required: [success, status, message, request_id]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 400
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: One or more fields are invalid
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: must be a valid email address
        request_id:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 150
        total_pages:
          type: integer
          example: 8

    # ── Auth ───────────────────────────────
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token TTL in seconds
          example: 900
        user:
          $ref: "#/components/schemas/UserResponse"
    OTPRequest:
      type: object
      required: [identifier, purpose]
      properties:
        identifier:
          type: string
          description: Email address or phone number
          example: "+237670000000"
        purpose:
          type: string
          enum: [login, registration, password_reset, verification]
    OTPVerifyRequest:
      type: object
      required: [identifier, code, purpose]
      properties:
        identifier:
          type: string
          example: "+237670000000"
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "482910"
        purpose:
          type: string
          enum: [login, registration, password_reset, verification]
    Enable2FAResponse:
      type: object
      properties:
        secret:
          type: string
          example: JBSWY3DPEHPK3PXP
        qr_code_url:
          type: string
          example: "otpauth://totp/UrbanSanctuary:john@example.com?secret=JBSWY3DPEHPK3PXP&issuer=UrbanSanctuary"
    TOTPCodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
    PasswordResetConfirm:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
          maxLength: 128

    # ── Users ──────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        avatar_url:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [guest, homeowner, admin]
        is_verified:
          type: boolean
        totp_enabled:
          type: boolean
        loyalty_points:
          type: integer
          example: 500
        created_at:
          type: string
          format: date-time
...
    CreatePropertyRequest:
      type: object
      required: [title, description, neighborhood, address, price_per_night, bedrooms, bathrooms, max_guests]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Luxury Studio in Bonapriso
        description:
          type: string
          minLength: 10
          example: A beautifully furnished studio apartment in the heart of Bonapriso with modern amenities.
        neighborhood:
          type: string
          enum: [akwa, bonapriso, bonanjo, kotto, other]
          example: bonapriso
        address:
          type: string
          example: 45 Rue de la Joie, Bonapriso
        latitude:
          type: number
          format: double
          example: 4.0483
        longitude:
          type: number
          format: double
          example: 9.6959
        price_per_night:
          type: number
          format: double
          minimum: 0
          example: 25000
        bedrooms:
          type: integer
          minimum: 1
          example: 1
        bathrooms:
          type: integer
          minimum: 1
          example: 1
        max_guests:
          type: integer
          minimum: 1
          example: 2
        amenities:
          type: array
          items:
            type: string
          example: [wifi, air_conditioning, parking, pool]
...
    PropertyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: string
          example: Luxury Studio in Bonapriso
        description:
          type: string
        neighborhood:
          type: string
          enum: [akwa, bonapriso, bonanjo, kotto, other]
        address:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        price_per_night:
          type: number
          format: double
          example: 25000
        bedrooms:
          type: integer
          example: 1
        bathrooms:
          type: integer
          example: 1
        max_guests:
          type: integer
          example: 2
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/PropertyImage"
        status:
          type: string
          enum: [draft, published, archived]
        average_rating:
          type: number
          format: double
          nullable: true
          example: 4.5
        review_count:
          type: integer
          nullable: true
          example: 12
        owner:
          $ref: "#/components/schemas/UserResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropertyImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
...